Refactor Cassandra / MySQL persistence layer * Add dedicated handling for WorkflowMutation and WorkflowSnapshot * Move Cassandra util code into dedicated file
